requirements.txt
# --- Ingestão e Processamento de Dados ---
yfinance
pandas
pandas-gbq
db-dtypes
google-cloud-bigquery
google-cloud-bigquery-storage 

# --- Analytics Engineering (dbt) ---
dbt-bigquery

# --- Camada de Inteligência (LangChain + Gemini) ---
google-generativeai 
langchain
langchain-google-genai>=1.0.10 
langchain-community
langchain-experimental
sqlalchemy-bigquery

# --- Utilitários e Infraestrutura ---
psycopg2-binary
python-dotenv